# tixmoapi2/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TixMo API",
    "version": "0.1.0",
    "description": "Enterprise-grade ticket vendor platform API for event promoters",
    "contact": {
      "name": "TixMo Team",
      "email": "api@tixmo.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.tixmo.com/v1",
      "description": "Production server (future)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token"
      }
    },
...
  },
...
  "paths": {
    "/analytics/sales": {
      "get": {
        "summary": "Get sales analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales analytics retrieved successfully"
          }
        }
      }
    },
    "/analytics/events": {
      "get": {
        "summary": "Get event analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event analytics retrieved successfully"
          }
        }
      }
    },
    "/analytics/customers": {
      "get": {
        "summary": "Get customer analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer analytics retrieved successfully"
          }
        }
      }
    },
    "/analytics/dashboard": {
      "get": {
        "summary": "Get dashboard summary (last 30 days)",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard summary retrieved successfully"
          }
        }
      }
    },
...
    "/events": {
      "post": {
        "summary": "Create a new event",
        "description": "Create a new event for an organization. Requires ADMIN or PROMOTER role.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "organizationId",
                  "venueId",
                  "startDateTime",
                  "endDateTime"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Summer Music Festival 2025"
                  },
                  "description": {
                    "type": "string",
                    "example": "Join us for an amazing night of live music"
                  },
                  "organizationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "venueId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "startDateTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-07-15T19:00:00Z"
                  },
                  "endDateTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-07-15T23:00:00Z"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PUBLISHED",
                      "CANCELLED"
                    ],
                    "default": "DRAFT"
                  },
                  "capacity": {
                    "type": "integer",
                    "example": 5000
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization or venue not found"
          }
        }
      },
      "get": {
        "summary": "List all events",
        "description": "Get a paginated list of events with optional filters",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "venueId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CANCELLED"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in title and description"
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}